Dev2 (#12)
* Moved entrypoints

* Added temporary policy

* Added Examples

* Changes env configuration

* Changes env configuration

* Added terrains

* Added instanced usd

* Adds USD

* adds debug terrain
