+++ rover_envs/assets/terrains/debug/__init__.py
@@ -1,1 +1,1 @@
-from .debug_terrains import DebugTerrainSceneCfg  # noqa F401
+# from .debug_terrains import DebugTerrainSceneCfg  # noqa F401
