+++ rover_envs/envs/navigation/__init__.py
@@ -1,6 +1,6 @@
-from .rover_env import RoverEnv
-from .rover_env_cfg import RoverEnvCfg
-from .rover_env_no_obstacles_cfg import RoverEnvNoObstaclesCfg
+# from .entrypoints.rover_env import RoverEnv
+# from .rover_env_cfg import RoverEnvCfg
+# from .rover_env_no_obstacles_cfg import RoverEnvNoObstaclesCfg
 
-__all__ = ["RoverEnvCfg", "RoverEnv", "RoverEnvNoObstaclesCfg"]
+# __all__ = ["RoverEnvCfg", "RoverEnv", "RoverEnvNoObstaclesCfg"]
 # __all__ = ["RoverEnvCfg", "RoverEnv", "RoverEnvCamera"]
