+++ rover_envs/envs/navigation/entrypoints/__init__.py
@@ -1,0 +1,1 @@
+from .rover_env import RoverEnv  # noqa F401
