+++ rover_envs/envs/navigation/rover_env_cfg.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import math
+from dataclasses import MISSING
 
 import omni.isaac.orbit.sim as sim_utils
 from omni.isaac.orbit.assets import ArticulationCfg, AssetBaseCfg
@@ -23,12 +24,9 @@ from omni.isaac.orbit.utils import configclass
 from omni.isaac.orbit.utils.noise import AdditiveUniformNoiseCfg as Unoise  # noqa: F401
 
 import rover_envs.envs.navigation.mdp as mdp
-from rover_envs.assets.aau_rover_simple import AAU_ROVER_SIMPLE_CFG
-from rover_envs.assets.terrains.debug import DebugTerrainSceneCfg  # noqa: F401
+from rover_envs.assets.terrains.debug.debug_terrains import DebugTerrainSceneCfg  # noqa: F401
 from rover_envs.assets.terrains.mars import MarsTerrainSceneCfg  # noqa: F401
 from rover_envs.envs.navigation.utils.terrains.terrain_importer import RoverTerrainImporter  # noqa: F401
-
-# from rover_envs.envs.navigation.utils.terrains.terrain_importer import TerrainBasedPositionCommandCustom
 
 ##
 # Scene Description
@@ -65,8 +63,9 @@ class RoverSceneCfg(DebugTerrainSceneCfg):
         init_state=AssetBaseCfg.InitialStateCfg(pos=(0.0, -180.0, 80.0)),
     )
 
-    robot: ArticulationCfg = AAU_ROVER_SIMPLE_CFG.replace(
-        prim_path="{ENV_REGEX_NS}/Robot")
+    robot: ArticulationCfg = MISSING
+    # AAU_ROVER_SIMPLE_CFG.replace(
+    #     prim_path="{ENV_REGEX_NS}/Robot")
 
     contact_sensor = ContactSensorCfg(
         prim_path="{ENV_REGEX_NS}/Robot/.*_(Drive|Steer|Boogie|Body)",
